# perf(recover): replace bytes with strings in function for better performance