@appleboy 2025-05-31
@1911860538 Can you provide the HTTP benchmark report using https://github.com/gin-gonic/go-web-framework-benchmark? Show the before and after results. Thanks.