@1911860538 2025-05-31
I used the load testing tool `hey` to compare the performance of CustomRecoveryWithWriter before and after my optimization.

## test code
```go
package main

import (
	"log"
	"net/http"

	"github.com/gin-gonic/gin"
)

func main() {
	gin.SetMode(gin.ReleaseMode)

	r := gin.New()

	r.Use(gin.CustomRecoveryWithWriter(gin.DefaultErrorWriter, func(c *gin.Context, err any) {
		c.JSON(http.StatusInternalServerError, gin.H{
			"message": "internal error",
		})
	}))

	r.GET("/panic", func(c *gin.Context) {
		panic("trigger panic")
	})

	err := r.Run(":8080")
	if err != nil {
		log.Fatal(err)
	}
}
```

## hey command
### hey install
```shell
brew install hey
```
### hey run
```shell
hey -n 100000 -c 100 http://localhost:8080/panic
```

## benchmark results before optimization using hey
```text
hey -n 100000 -c 100 http://localhost:8080/panic

Summary:
  Total:        26.3267 secs
  Slowest:      0.3569 secs
  Fastest:      0.0005 secs
  Average:      0.0256 secs
  Requests/sec: 3798.4205

  Total data:   3200000 bytes
  Size/request: 32 bytes

Response time histogram:
  0.001 [1]     |
  0.036 [71640] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.072 [18683] |■■■■■■■■■■
  0.107 [6305]  |■■■■
  0.143 [2242]  |■
  0.179 [755]   |
  0.214 [257]   |
  0.250 [78]    |
  0.286 [26]    |
  0.321 [9]     |
  0.357 [4]     |


Latency distribution:
  10% in 0.0013 secs
  25% in 0.0025 secs
  50% in 0.0061 secs
  75% in 0.0393 secs
  90% in 0.0707 secs
  95% in 0.0945 secs
  99% in 0.1465 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0000 secs, 0.0005 secs, 0.3569 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0083 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0106 secs
  resp wait:    0.0256 secs, 0.0005 secs, 0.3568 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0146 secs

Status code distribution:
  [500] 100000 responses
```

## benchmark results after optimization using hey
```text
hey -n 100000 -c 100 http://localhost:8080/panic

Summary:
  Total:        25.3953 secs
  Slowest:      0.4092 secs
  Fastest:      0.0006 secs
  Average:      0.0246 secs
  Requests/sec: 3937.7382

  Total data:   3200000 bytes
  Size/request: 32 bytes

Response time histogram:
  0.001 [1]     |
  0.041 [78627] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.082 [14995] |■■■■■■■■
  0.123 [4648]  |■■
  0.164 [1289]  |■
  0.205 [328]   |
  0.246 [82]    |
  0.287 [23]    |
  0.327 [5]     |
  0.368 [1]     |
  0.409 [1]     |


Latency distribution:
  10% in 0.0013 secs
  25% in 0.0025 secs
  50% in 0.0059 secs
  75% in 0.0376 secs
  90% in 0.0682 secs
  95% in 0.0901 secs
  99% in 0.1394 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0000 secs, 0.0006 secs, 0.4092 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0060 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0047 secs
  resp wait:    0.0245 secs, 0.0005 secs, 0.4091 secs
  resp read:    0.0000 secs, 0.0000 secs, 0.0170 secs

Status code distribution:
  [500] 100000 responses
```